requirements.txt
# FastAPI + Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0
python-multipart>=0.0.9

# LLM (direct Anthropic SDK — no LangChain overhead)
anthropic>=0.40.0
httpx>=0.27

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9

# Data processing
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0

# PMS pipeline
openpyxl>=3.1

# Optional — install separately if RAG features needed:
# pgvector>=0.3.0
# sentence-transformers>=3.0
# pymupdf>=1.24
